old2/requirements.txt
pydantic
pydantic-settings
pymongo
pyyaml
python-dotenv
python-telegram-bot